Fix ack processing when acked interval partially covers segment
Because of gso, an ack_seq may be transmitted that falls in the middle of a
captured segment. In that case, there will never be an ack_seq update
interval that fully includes that segment and it will linger in the
unacked_q.

Mark a segment as acked once its end_seq is acked.